move inference out of Base into Core.Inference
having inference inside Base made it hard to modify and
slow to compile the sysimg since bootstrapping.
additionally, it made inference vulnerable to unintentional
sabotage by the users program.

to combat those issues, this creates a copy of the essential code
required for running the inference into Core.Inference.
this allows any part of the system (esp. Base) around inference to
change, and even be recompiled or monkey-patched,
without impacting the core inference algorithms.

note that this clarifies and even tightens the restrictions on what
functionality is available for the use of inference.
the primary available functionality is:
  "essential" primitives
  integers
  arrays
  dictionary support (limited, and ObjectIdDict will NOT deserialize properly)
  c-interaction
  intset
  limited printing and string support (no interpolation)